Conflicts:
	src/main/org/nlogo/hubnet/mirroring/Agent.java
	src/main/org/nlogo/hubnet/mirroring/AgentData.java
	src/main/org/nlogo/hubnet/mirroring/AgentPerspective.java
	src/main/org/nlogo/hubnet/mirroring/ClientWorld.java
	src/main/org/nlogo/hubnet/mirroring/DiffBuffer.java
	src/main/org/nlogo/hubnet/mirroring/HubNetDrawingMessage.java
	src/main/org/nlogo/hubnet/mirroring/HubNetLine.java
	src/main/org/nlogo/hubnet/mirroring/HubNetLinkStamp.java
	src/main/org/nlogo/hubnet/mirroring/HubNetPlotPoint.java
	src/main/org/nlogo/hubnet/mirroring/HubNetTurtleStamp.java
	src/main/org/nlogo/hubnet/mirroring/LinkData.java
	src/main/org/nlogo/hubnet/mirroring/Overridable.java
	src/main/org/nlogo/hubnet/mirroring/OverrideList.java
	src/main/org/nlogo/hubnet/mirroring/PatchData.java
	src/main/org/nlogo/hubnet/mirroring/ServerWorld.java
	src/main/org/nlogo/hubnet/mirroring/TurtleData.java

+++ src/main/org/nlogo/lab/ProtocolLoader.scala
@@ -1,4 +1,4 @@
-// (C) 2012 Uri Wilensky. https://github.com/NetLogo/NetLogo
+// (C) Uri Wilensky. https://github.com/NetLogo/NetLogo
 
 package org.nlogo.lab
 
@@ -73,7 +73,7 @@ class ProtocolLoader(services: CompilerServices)
         !element.getElementsByTagName(name).isEmpty
       def valueSets = {
         def readSteppedValueSetElement(e: dom.Element) = {
-          def parse(name: String) = e.getAttribute(name).toDouble
+          def parse(name: String) = BigDecimal(e.getAttribute(name))
           new SteppedValueSet(e.getAttribute("variable"),parse("first"),
                               parse("step"),parse("last"))
         }
